+++ sky/provision/kubernetes/utils.py
@@ -1139,3 +1139,14 @@ def combine_pod_config_fields(config_yaml_path: str) -> None:
 
     # Write the updated YAML back to the file
     common_utils.dump_yaml(config_yaml_path, yaml_obj)
+
+
+def check_nvidia_runtime_class() -> bool:
+    """Checks if the 'nvidia' RuntimeClass exists in the cluster"""
+    # Fetch the list of available RuntimeClasses
+    runtime_classes = kubernetes.node_api().list_runtime_class()
+
+    # Check if 'nvidia' RuntimeClass exists
+    nvidia_exists = any(
+        rc.metadata.name == 'nvidia' for rc in runtime_classes.items)
+    return nvidia_exists
